Updated openstack/openstack Project: openstack/cinder  dfb69c1163bd91c97015b365bec0963a3f24ecf5
Replace urllib.unquote with urllib.parse.unquote

This patch replaces urllib.unquote with urllib.parse.unquote
in the vmdk driver.

Commit d08c7ffe52e3ee211e425c363200998704726e58 replaced
urllib and urllib2 with six.moves.urllib and changed urllib.
unquote to urllib.parse.unquote but missed one call site
in the vmdk driver since the code which introduced it was
under review at the same time. Due to this, if we set an
optional parameter 'vmware_cluster_name', the driver init
will fail with AttributeError.

This patch also adds a unit test which would have caught
this error during gate check.

Change-Id: I310456f5d677c6e100a91efc6228c533cb5bf208
Closes-Bug: #1488916