Updated openstack/openstack Project: openstack/nova  19810750b41a063262bdb6a6f57e759a1bf951df
Pre-load expected attrs that the view builder needs for server details

There is a race between getting the instances from the database and
when the view builder shows the details.  If an instance is deleted
before lazy-loading some info, like the flavor, then the view
builder fails with an InstanceNotFound and the request returns a 500
in the v2.1 case and a 404 in the v2 case.

We can have the view builder tell the API code what attributes it
expects when it gets the instance from the database so that we don't
have to do any lazy loading in the view builder.

Change-Id: Ie03ab415baaa9e88c423a1c60a40d53b4cc464c4
Closes-Bug: #1487570