Updated openstack/openstack Project: openstack/horizon  47c8a18e7822a168b2637da9682e1db8c0d6510f
Revert "Fix broken header dropdowns"

This reverts commit 59a183aed82552462bb814fc13abb1f0217bca03.

With 0.11 this causes every <a data-toggle="dropdown"></a> to 
require 2 clicks for opening - first click selects it, second opens.

There is not even a package for XStatic-Angular-Bootstrap 0.13 yet. 
This patch either needs to be compatible with 0.11 and 0.13, or we 
need a new 0.13 package and update global-requirements lower end 
cap on it.

Change-Id: I853f14a813b68faa57b3751fff8486e37c3014a5