Updated openstack/openstack Project: openstack/nova  8124bdec874ce02e1a1f513c24534ddc2198932a
libvirt: take account of disks in migration data size

Currently migration is tuned based on the guest RAM size
alone, but when doing block migration we must also look
at the guest disk size to determine total data transfer.

This change currently follows the (stupid) libvirt logic
for deciding which disks are copied. In the future when
we can use new libvirt which lets us specify a desired
list of disks we'll update this logic.

DocImpact: the amount of time a live migration can take
is now configurable for the libvirt driver, and is
based on a number of seconds per gigabyte of RAM and
local disk.

Closes-bug: #1429220
Change-Id: I3d525e62c6686277c6a05f2a91edefad3230c73f