Updated openstack/openstack Project: openstack/oslo.vmware  489f9ec1ac31689e42ddcf66c26fded6a722c5df
Use SSL thumbprints for NFC transfer

There are cases when clients cannot build a trust chain in order to
verify ESX certificate. The SSL thumbprint is always available and it is
the preferred way for verifying the connection when doing NFC transfer.
This patch replace the usage of truststore with SSL thumbprints in
VmdkReadHandle and VmdkWriteHandle.

Change-Id: I9573bada15aef2bf8d25fa824bbbc2b421586733