Updated openstack/openstack Project: openstack-infra/puppet-cgit  dc7e58943dbbb4283950f8009623d6ae8ec7f9f2
Use systemd'd git-daemon on Centos7

Centos7 is a bit more opinionated on how git-daemon should run. In
particular with selinux the git_system_t context does not have
permissions to the git_port_t port(s) because systemd is expected to do
socket activation for git-daemon.

Fix this by not fighting systemd and embracing it. Use it for socket
activation with the git-daemon process and potentially add the git
daemon port to git_port_t label if necessary.

Change-Id: Id3fadfa74261649d158f4f31879f74f83d5856a8