Updated openstack/openstack Project: openstack/diskimage-builder  5bae00db0cbd19fc953b199978e3d45db81b40c2
Make README.rst a bit more generic

I think diskimage-builder has grown somewhat since this was written,
and this makes it sound like it is only useful as a TripleO component.

Give a more generic overview, add a simple example to illustrate the
point and generally make it clear this is a flexible solution for
building disk images.  Add a mention of the local documentation
source.

Change-Id: I74f13ccc1e3fe50ac907efcc7f479e59fe44455c