Updated openstack/openstack Project: openstack/nova  dc6b6559fdc9ca7b42c1015d3b9f94d383a112d0
Xen: import migrated ephemeral disk based on previous size

When migrating an instance with the xenapi driver it copies the VHDs
over to the destination host and then imports them into the SR on that
host.  It was attempting to import ephemeral disks based on the size of
the ephemeral disk for the new flavor, not the old flavor being
migrated.  This caused a failure when the old flavor did not have an
ephemeral and the new flavor did so Nova was attempting to import an
ephemeral disk that didn't exist.

Now ephemeral disks will be imported based on what the old flavor would
have migrated.

Change-Id: I80485fd473e8b580cc4b0a537c716bf6202a5eb2
Closes-Bug: 1488635