Updated openstack/openstack Project: openstack/nova  58e17ca79d2cad088b4c78dd3533c6fa937aa41e
Add a note about the 403 response not requiring a microversion

Change Ia782ab9d9a31d37854606f7be82582de6d7bc374 returns a 403 rather
than a 500 for OverQuota and adds that to the list of expected errors in
the v2.1 case.

There is no microversion change required since the 403 case is implied
given keystone authentication can fail with a 403 on any API.

This change updates the docs to mention the exceptional case for 403.

Related-Bug: #1444559

Change-Id: I7ea8c5c50d4eac53837f5f76e09277ac185288a9