Updated openstack/openstack Project: openstack/nova  1df358503059382d8a0e0eb66b19521963f9dde5
Convert percent metrics back into the [0, 1] range

With the recent revamp of the Nova metrics framework in the Liberty
cycle, there was a problem introduced in terms of how the percentage-
based metrics (e.g., cpu.percent) are represented. That is, prior to
Liberty, we stored the percentages in compute_nodes.metrics as a
floating point value in the range [0, 1]. For example, 17.5% was
represented as 0.175.

The revamped framework expects the MonitorMetric.value properties to
be integers, so when the percentages such as 0.175 are stored in this
object, they're erroneously converted to 0.

This patch multiplies the 'problematic' fields by 100 in preparation
for storage within the MonitorMetric object and then divides the
same problematic fields by 100.0 when converting back to their JSON
format, which will preserve RPC notifier and database compatibility.

Added some unit tests to test the conversions as well as making some
of the unit tests more explicit so this doesn't happen again.

Change-Id: I17a9412c63eeeb78ba2b2600f9007e9b58fcbf68
Closes-Bug: #1488696