Updated openstack/openstack Project: openstack/ironic  5754ab65ced6527dcef55cccda348d12efe5bbe1
Put py34 first in the env order of tox

To solve the problem of "db type could not be determined" on py34
we have to run first the py34 env to, then, run py27. This patch
puts py34 first on the tox.ini list of envs to avoid this problem
to happen.

Change-Id: Ib09cbc71149ffcc1f36495faf1c6586c4e0a0a86
Closes-bug: #1489059