Updated openstack/openstack Project: openstack/neutron  c318db29373467a9eb6696baf52748616782556b
devref: added guidelines on how to maintain sub-projects

I was involved lately in several boring attempts to fix broken
sub-project requirements and stable branches, and I think part of the
problem here is that we don't communicate our expectations to
sub-projects clear enough.

This is a first attempt to set brief and clear guidelines to sub-project
maintainers on how to maintain their repositories without much hassle.

Change-Id: I9180ee530f96a885b5667e050d141ce0ab52a8ce