Updated openstack/openstack Project: openstack-infra/release-tools  e11f0e8f50ac91de72fda2b61ec76ff27dc5d7b5
Fix formatting in process_bugs.py output

The recent changes to use print() instead of print mean we started
having more newlines, which in turn means the output is much longer than
it used to be. Restore the same formatting, and fix some prints that
didn't have the () so weren't being called.

Change-Id: I55f8c8bfe899bc0bb57cdf4f4a4b49ffd146ff94