Updated openstack/openstack Project: openstack/api-site  27ccaa0f618de4231ec257d0465fae915050c080
Fix the URI parameter of 'List shared images' in Image service API v1

The original URI of this API is '/v1/shared-images/{owner}'.
But there is no specific description of the parameter 'owner'.
Meanwhile the other APIs in the same page are using 'owner_id'
instead of 'owner'.

So the following changes are made.
 - Change the URI from '/v1/shared-images/{owner}' to '/v1/shared-images/{owner_id}'.
 - Add the description of the parameter 'owner_id' as request parameter.
 - Move the image_id, owner_id and can_share parameters to common.ent
   and reference those entities from the WADL.
 - Add return at end of code samples - those were missing.

Change-Id: I7b178f619266ddd93af48921c16561831bf1287c
Closes-Bug: #1489383