Updated openstack/openstack Project: openstack/oslo.log  f718919c459be88d24fbf96a97856835c611b7a8
Fix poor examples of exception logging

There were several examples showing passing an
exception in as the message text to a call to
LOG.exception. LOG.exception will automatically
log the exception details, so doing something
like this results in the exception information
being written out twice.

To prevent the possibility of someone seeing these
examples and assuming that is the correct way to
use LOG.exception, examples have been updated to
use LOG.error instead to illustrate the correct
way to pass in exceptions in the message string.

Change-Id: I582008609180a02eaff3d85bee5b5ca4381719ce