Updated openstack/openstack Project: openstack/barbican  7899f9a5314f643c270272d4ca4436777798f84f
Fixed Testing Attributes Bug

There was a bug in barbican/tests/utils.py when building parameterized
tests. Attribute tags were unenforced when used in conjunction with a
parameterized test. This patch fixes that bug. The issue was that
attributes were assigned to the function but not transferred to the
newly created parameterized functions.

Change-Id: I031d419b366dee48dc4dd9b8150b3f83d81678d5
Closes-Bug: #1489900