Updated openstack/openstack Project: openstack-infra/project-config  dc2412ed26a51a755d0ad109ce67c6f2922055f2
Remove Neutron functional job from the gate queue

The job has had a high failure rate for the past month:
http://goo.gl/oTfn69

As a result of a group of bugs:
https://goo.gl/oEpmQd

It's time to knock it off the gate queue. It would be preferable
to keep it in the Neutron check queue as voting, therefor remaining
an incentive to lower the failure rate for the Neutron community.

Change-Id: I1e3a2cccea61deec3d746d43509bc85914ca72ef