Updated openstack/openstack Project: openstack/cinder  dd3b25fbff1e4d74d2d7d5cf28712e1f0e40c4a4
Rework Scality SOFS driver to use RemoteFS class

This change is big but the logic is simple. The previous version of
this driver reimplemented most of what is now in the RemoteFS base
class. SOFS stands for Scale Out FileSystem and is based on FUSE, so
it's only natural to leverage the RemoteFS class.

Change-Id: I26935061e860a477f98a61a935e67281a3a6f48e