Updated openstack/openstack Project: openstack/heat  1fab48d2664094c91dda7677f890812115592584
Use latest stable template version in snippets

2015-04-30 is the latest template version in a stable heat
release, so it should be used for resource reference snippets.

While there is potential for keeping resource.py on the latest
stable using python, this approach just replaces the snippet
string so it can be simply updated at the same time as the rst
files.

Change-Id: I95802fa7a6853d52a7e78f420853ef2a88d6c5e4