Updated openstack/openstack Project: openstack/nova  be2db9b839d43b97ec02c8efc874712c22195c66
Handle neutron exception on bad floating ip create request

This commit adds a new nova exception to handle a floating ip create
failure with a BadRequest exception from neutronclient. This is then
used in all 3 nova api implementations to ensure a 400 response is
returned just as neutron is returning to nova.

Change-Id: I36e8ef0113ae91e3b15c846c6d10e0b766cf2a37
Closes-Bug: #1482816