Updated openstack/openstack Project: openstack/neutron-fwaas  d10ef329d69ae8f2e2df02d64a5099821a47ee49
FWaaS: Fix 500 INTERNAL error on too long name or description

The firewall interface was not validating string lengths, so
passing in too long of a name (>255 char) or description (>1024)
on a create returns a 500 INTERNAL SERVER error from a failed
insert at the DB level.

Solution is to validate the name and description strings in the
firewall interfaces so a invalid input length user error returned
instead.  Used the standard attributes NAME_MAX_LEN and
DESCRIPTION_MAX_LEN commonly used in other neutron interfaces
for same validation.

Fixes the 'firewall-create', 'firewall-policy-create' and
'firewall-rule-create' for both name and description variants.

Added unit test cases for long names and descriptions.
Merged with REJECT updates.

Change-Id: I188140ae5e9624d2cc35ae980fefe35659ddaae2
Closes-Bug: 1489125