Updated openstack/openstack Project: openstack/ironic  55283a9cfa93e925592297aee17c8fed77d338a9
Periodically checks for nodes being cleaned

This patch is adding a periodic task to check for nodes waiting for the
ramdisk callback when cleaning is being executed.

A new configuration option called "clean_callback_timeout" was added,
its value is the number of seconds that the Ironic conductor will wait
for the ramdisk to doing the cleaning to contact Ironic back. Defaults
to 1800.

Closes-Bug: #1483120
Change-Id: Id7f9e9018b5cb2389bbe556171e7a9d46425afba