Updated openstack/openstack Project: openstack/nova  e02b976846734604844ecd9d179d9cb661b8cd15
Allow for migration object to be passed to _move_claim

In case of evacuate, we want to be able to create a migration record as
early as possible, but still be able to do the claim for resoruce
tracking only once the request is actually being claimed on the
destination compute node.

For this we'll need to pass in the already created Migration, so this
commit adds this as an argument to the _move_claim() method and adds the
relevant code paths.

Related-blueprint: migration-fix-resource-tracking
Change-Id: I295e83c72b880acce2a64f87c98f02ff4aff79e8