Updated openstack/openstack Project: openstack/horizon  9ee1106472e6ad88b49d11a1038613a921462276
Styles and Asset No Longer Necessary

There are two styles in horizon.scss that look like they are no longer
used. In addition, one of the rules is pulling in an asset,
right_droparrow.png that is only used in this style. It looks like all
of these can be removed.

Change-Id: I6fd5cde9f7e31565864913e76d0ea1049e82af7f
closes-bug: #1489703