Updated openstack/openstack Project: openstack/nova  938a20f5e29ec73ea01089406a8624f6d4e9e093
Use compatibility methods from oslo

This removes the nova mirrors of the compatibility and primitive
methods from oslo.versionedobjects. We needed to keep these until we
got on oslo's registry. We've done that, and now we need to drop these
so that we can take advantage of new features in oslo.

Related to bp/use-oslo-objects

Change-Id: I01aa967f524f80a2dfadee328cdbf55d0b96466b