Updated openstack/openstack Project: openstack/nova  5099bd5d1a6afff70eeacfdaf18d8ee9a7b26f71
Fix IronicHostManager to skip get_by_host() call.

The recent change to remove the direct nova database calls by several
scheduler filters added a system to have the compute hosts update the
scheduler's view of their instances on a regular basis, with an option
to opt out. However, the opt out required that the Host Manager run a DB
query for every host on every request, whether the information was
needed for the filters or not. This has turned out to be very
problematic for ironic, whose hosts can have hundreds or thousands of
instances.

This patch provides a quick fix by overriding the _add_instance_info()
method to simply set the host_state.instances value to an empty dict. It
also overrides the _init_instance_info() method to a no-op, since that
will never be used in ironic.

Co-Authored-By: Ed Leafe <ed@leafe.com>
Change-Id: I18075fd2fd8f9f6d1d82e6cc487c9e15ebc2b1d5
Closes-Bug: #1479124