Updated openstack/openstack Project: openstack/neutron-fwaas  aabf6d42a0f07ce3e9be8ce888fffb2fc88509be
Kill contention between update and delete

Patch I455bbe00395377509beca5ec1d54cf4f7a57152d didn't go enough and we knew it. This patch
overhauls the status update logic to kill the race contention once and for all by updating in
place (or at least that's the wishful thinking).

The logic of preventing updates for firewall in progress of being deleted is preserved. This
is done by introducing a new update method that encapsulates DB logic where it belongs.

Code being affected is covered already so no more coverage is deemed necessary.

Closes-bug: 1483875

Change-Id: Ie6fc6d2545b778f8264d57201827dd82618259f9