Updated openstack/openstack Project: openstack/nova  4ed4546892b61e989903e5897840eed24e20997e
downgrade log messages for memcache server (dis)connect events

The connect and disconnect events for the service group memcache
server were being logged at ERROR level, this changes them to INFO for
connect and WARN for disconnect.

It does not narrow the exception catch because the class of errors
this can expose are not clear, and we don't test any of this in full
stack testing.

Change-Id: I5dd96cf90a9ffa3c05fee9e7a8e2c70fc6ac1140