Updated openstack/openstack Project: openstack/trove  2e621d48d67cb5fa0ce7f7fc17d2f694fff5c02f
MySQL Manager Refactor

Creates a class structure for MySQL-derived datastores to avoid
duplication of code for features and capabilities shared in common.

The existing manager and service classes were pulled up into the
manager_base and service_base modules. Module constants in the manager
and service modules were changed to class properties, so that the
various implementations can inject the appropriate dependencies.
The original logic was preserved to the extent possible.

Change-Id: I5217390c9ff0cdb1b781dd8f2291cca23f4e442c
Implements: blueprint mysql-manager-refactor