Updated openstack/openstack Project: openstack/nova  1a9708a6c01bc22f409612a47337127583944473
Fix collection of metadata for a snapshot of a volume-backed instance

Currently the snapshot derives its properties from the instance source:
an other snapshot or a bootable volume. But that sources could be
changed since the instance was booted.

To make instance snapshots independent of source changes this patch
collects metadata from instance system metadata rather than the sources.

Since it brings the only way to initialize image metadata, it fixes:
a) min_ram attribute is not inherited from a bootable volume
(LP #1369854).
b) empty container_format and disk_format attribute are inherited from a
source instance snapshot (LP #1439819).

Closes-Bug: #1369854
Closes-Bug: #1439819
Related-Bug: #1469179
Change-Id: I067f66356a5ebd738add1591a0069d8049f35c24