Updated openstack/openstack Project: openstack/nova  1d0f5b0a98e9a54b3edce8485f03bd5ca8535e52
Consolidate initialization of instance snapshot metadata

Both types of instance snapshots (volume-backed and image-based) require
similar metadata with a few differences. This patch introduces a common
method for initialization of instance snapshot metadata and uses it in
the custom cases.

Also this fixes a bug for volume-backed instance snapshots: properties
restricted to be inherited by config are inherited though.

Closes-Bug: #1384392
Change-Id: I811405bbe98e391f76b868cd35b03ca3dced8cdd