Updated openstack/openstack Project: openstack/neutron  cc49b7d7d9e8ba32cb7f4e9635ea20a7b7266b9b
IPv6 display suitable message when MTU is invalid on iface

IPv6 protocol requires a minimum MTU of 1280 bytes on the interface
to configure an IPv6 address to the interface. This patch logs an
appropriate error message and exits the agent if ipv6 is enabled and
network_device_mtu is less than the minimum value.

DocImpact

Closes-Bug: #1475015
Change-Id: I13666de4e6f5f6775ad26342e513c3c17a003b8e