Updated openstack/openstack Project: openstack/openstack-manuals  b3d592ccf1ee97c8471e7da4129868b6b2a744dd
Fixed manual [OpenStack End User Guide],section[Launch and manage stacks]
original description is as following:
---------------------------------------------------------------------------------------
Depending on the source that you select, enter the URL, browse to the file location, or
directly include the the template.
---------------------------------------------------------------------------------------
I think "include the the template" should be modified to "include the template".
Closes-Bug: #1489774

Change-Id: I5df2abd5cfba5259025400a17b4e24e90c8411b0