Updated openstack/openstack Project: openstack/openstack-manuals  a328532fbbc61a4aa13e2fdceaf65b7489547334
Repeated words in Networking Guide

The origninal description and my changes are as following:

PCIDeviceScheduler to the the scheduler_default_filters parameter

I changed " the the " to "the ".

Edit the the 
/etc/neutron/plugins/ml2/ml2_conf.ini file:

I changed " the the " to "the ".

The The ``qr-2`` interface contains the project network 2

I changed " The The " to "The ".

Updated all RST files.

Closes-Bug: #1489682
Change-Id: I62b106c719bdade3660eb39783cc0e64cf302fb1