Updated openstack/openstack Project: openstack-infra/project-config  31cad869a7b16e0a9bfbc45bef2235d54dcb5f14
Splitting stackforge/freezer on multiple repos

The main freezer repo now contains the API and Web UI
code. As the project is growing we would like to have
an individual repo for each component.

This commit refer to the creation of a new freezer-web-ui repo.
The directory horizon_web_ui of the stackforge/freezer repo
needs to be imported on this new repo with the history.

Implements blueprint: main-repo-split

Change-Id: I2e7ea83247eec090f404f66575c9d7b8e19fa894