Updated openstack/openstack Project: openstack/nova  c7d355c6a6e6ede4f357277f4d941f5185fd3863
api: deprecate the concept of extensions in v2.1

We want to remove the extension infrastructure in the v2.1 API during
the next release. To make that possible, we must deprecate the
configuration now.

Also deprecating the osapi_v3 group, and moving to osapi_v21.

UpgradeImpact
DocImpact

Part of blueprint nova-api-deprecate-extensions

Change-Id: I084444b11dceda7cf8f88c157aa67d36490fce49