Updated openstack/openstack Project: openstack/nova  51a6665389fa22fc8646789fc25400a9a943e772
Consider that all scheduler calls are IO Ops

There was a conditional checking either the vm_state or the task_state to see
if it was an IO operation. By thinking on it, there are 5 calls to the scheduler
for asking a destination :
 - build_instances
 - cold_migrate
 - live_migrate
 - unshelve
 - rebuild

All of them will try to find a destination and consume_from_instance() will be
decrementing the host resources by the instance usage, which means it's an I/O
operation anyway.

Also, the current implementation is buggy since a live migration is not
considered as an io_ops (the MIGRATING task_state is missing in the conditional...)
so that's a good opportunity for just saying bye-bye to the conditional.

Partially-Implements: blueprint request-spec-object

Change-Id: I7ca1b887bbb3a77e0cbb657e39926e9e3e393df8