Updated openstack/openstack Project: openstack/nova  adfb621cfe5933ad774208ad711185653648650d
Cleanup RT _instance_in_resize_state()

Move the ResourceTracker._instance_in_resize_state() method to a simple
module-level function, since nothing in the method actually accessed the
resource tracker object anyway, and testing a simple function means we
can test it without the hassle of creating the resource tracker in the
test cases.

Also removes a duplicate test_claim() method from the
test_resource_tracker module that was already in the NoDBTestCase-based
test_tracker.MoveClaimTestCase object. The idea behind this and the next
patches are to reduce the amount of test code changes required when
changing the PciDeviceTracker's constructor and its instantiation in the
resource tracker.

Change-Id: I0e0aebbe6ca0f659068e1adb2c9a1d6231fb70a7