Updated openstack/openstack Project: openstack/nova  ae49a408cbb2c4b5228b987b5ef091c5243f723b
xapi: ensure pv driver info is present prior to live-migration

Xenserver relies on guest (domU) to provide information about
the presence of PV drivers in the guest image back to dom0 though
xenstore.

It is possible for users to disable the xen agent that reports this
info and therefore causing failures in live-migration. The agent is
only responsible for reporting guest infomation to dom0, and is not
directly responsible for any live-migration activity. Therefore, in
cases where PV drivers are running it is safe to fake the presence of
this information in xenstore. This plugin checks for such cases and
writes a fake PV tools information prior to live-migration.

The information about the presence of pv driver or version is read
straight though xenstore and written back to xenstore.

This information is checked by xapi (not nova) during live migration.

Change-Id: Ib9624a6f325516b0124ebdecebb0951f29ba3ac4
Closes-Bug: #1462295