Updated openstack/openstack Project: openstack/ironic  ec461e7e38cdfe9b59e5b374abf44902784fd9a0
Do not overwrite the iPXE boot script on every deployment

This patch just add a check to verify if the iPXE boot script have already
been laid down in the HTTP root directory and do not overwrite it if
it's already there. The script is the same for all deployments so we
don't need to write it over and over. This also makes it easier to test
some custom parameter in the script without having to deal with Ironic
undoing the changes prior to deploy a node.

Closes-Bug: #1489904
Change-Id: I0d56767978a6d1290442e50d7cb1308d78e3a352