Updated openstack/openstack Project: openstack/ironic-python-agent  a1f8db13cdadcd922ff28eb856b70d5a0e32d336
Load Hardware Managers at runtime

Hardware managers should load at runtime. This will ensure the agent is
ready to respond to API calls before it begins heartbeating. Also, it
means in case of a syntax or other error in a HardwareManager, the agent
will crash before it heartbeats, which is better than it working until a
hardware manager method is needed.

Change-Id: I9403ce7bedc8d5af20b6d84371367253b26b74c2
Closes-bug: 1490008