Updated openstack/openstack Project: openstack/nova  10880b9752339acb6aab02af9bfa45290650b40f
Add cpu_allocation_ratio and ram_allocation_ratio to ComputeNode

As we want to provide a per-compute allocation ratio for both CPU and RAM,
we need to persist them to the compute_nodes table and add them to the
ComputeNode object.
Note: Since we need to provide a compatibility path for upgrades and as we
don't want to ask the operators to change their config files before upgrading,
we then verify the allocation ratio value to match if defaulted and then return
an hardcoded value identitical to the behaviour we want.
In order to know if the compute related opt values were updated, we plan to
change the ratios defaults with 0.0 in a later change
(I382421968d4c7989a6ff08c74c21139bf6be12e5)
hence the if branch in the object.

Change-Id: I83074644c35e151927c23f6d7d50c9dc32e9bd2d
Partially-Implements: blueprint allocation-ratio-to-resource-tracker